Same for pipeline run links
